requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pandas
python-dotenv
sqlalchemy
psycopg2-binary
azure-ai-projects 
azure-identity 
opentelemetry-sdk 
azure-monitor-opentelemetry